pyproject.toml
[tool.poetry]
name = "projeto2"
version = "0.1.0"
description = ""
authors = ["arrudaguilherme"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = "^0.29.0"
sqlalchemy = "^2.0.29"
psycopg2-binary = "^2.9.9"
pytest = "^8.1.1"
alembic = "^1.13.1"
python-decouple = "^3.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
